NGC-427 Refactor introduced Money to clean up rounding and scaling

+++ app/uk/gov/hmrc/taxcalc/controllers/TaxCalculatorController.scala
@@ -26,7 +26,7 @@ import scala.concurrent.ExecutionContext.Implicits.global
 trait TaxCalculatorController extends BaseController with ErrorHandling {
   val service: TaxCalculatorService
 
-  def calculateTax(isStatePensionAge: Boolean, taxYear: Int, taxCode: String, grossPay: BigDecimal, payPeriod: String,  journeyId: Option[String]) = Action.async {
+  def calculateTax(isStatePensionAge: Boolean, taxYear: Int, taxCode: String, grossPay: Long, payPeriod: String,  journeyId: Option[String]) = Action.async {
     implicit request => {
       service.calculateTax(isStatePensionAge, taxYear, taxCode, grossPay, payPeriod).map{
         response => Ok(Json.toJson(response))
